Refactor validation utilities into new module
